+++ tests/Unit/Bootstrap4/Inputs/Label/InputLabelTest.php
@@ -10,9 +10,10 @@ class InputLabelTest extends \Okipa\LaravelFormComponents\Tests\Unit\Bootstrap5\
     public function it_can_set_input_file_label_class(): void
     {
         $html = $this->renderComponent(Input::class, ['name' => 'first_name', 'type' => 'file']);
-        self::assertStringContainsString('<label for="file-first-name" class="custom-file-label">', $html);
-        $labelPosition = strrpos($html, '<label');
-        $inputPosition = strrpos($html, '<input');
-        self::assertLessThan($labelPosition, $inputPosition);
+        $this->assertSeeHtmlInOrder($html, [
+            '<label for="file-first-name" class="form-label">',
+            '<input',
+            '<label for="file-first-name" class="custom-file-label">',
+        ]);
     }
 }